Move Configs to Const
