+++ eqa/const/config_newline.py
@@ -1,0 +1,17 @@
+NEW_LINE_OTHER_CONFIG = """
+{
+  "line": {
+    "all": {
+      "alert": {},
+      "reaction": false,
+      "sound": false
+    },
+    "undetermined": {
+      "alert": {},
+      "reaction": false,
+      "sound": false
+    }
+  },
+  "version": "%s"
+}
+"""
