+++ eqa/const/validspells.py
